+++ test/get-supported-languages.js
@@ -1,28 +1,26 @@
-var getSupportedLanguages = require('..').getSupportedLanguages
-var assert = require('assert')
-var fs = require('fs')
-var path = require('path')
+const { getSupportedLanguages } = require('..')
+const assert = require('assert')
+const fs = require('fs')
+const path = require('path')
 
-describe('getSupportedLanguages', function () {
-  it('lists all supported languages', function (done) {
-    var definitionsPath = path.resolve(__dirname, 'definitions')
+describe('getSupportedLanguages', () => {
+  it('lists all supported languages', done => {
+    const definitionsPath = path.resolve(__dirname, 'definitions')
 
-    fs.readdir(definitionsPath, function (err, files) {
-      if (err) { throw err }
+    fs.readdir(definitionsPath, (err, files) => {
+      if (err) { return done(err) }
 
-      var languages = files.filter(function (file) {
-        return path.extname(file) === '.csv'
-      }).map(function (file) {
-        return path.basename(file, '.csv')
-      })
+      const languages = files
+        .filter(file => path.extname(file) === '.csv')
+        .map(file => path.basename(file, '.csv'))
 
-      assert.deepEqual(languages.sort(), getSupportedLanguages().sort())
+      assert.deepStrictEqual(languages.sort(), getSupportedLanguages().sort())
 
       done()
     })
   })
 
   it('returns a different array each time', function () {
-    assert.notEqual(getSupportedLanguages(), getSupportedLanguages())
+    assert.notStrictEqual(getSupportedLanguages(), getSupportedLanguages())
   })
 })