+++ test/package.js
@@ -1,14 +1,14 @@
-var pkg = require('../package.json')
-var bower = require('../bower.json')
-var assert = require('assert')
+const pkg = require('../package.json')
+const bower = require('../bower.json')
+const assert = require('assert')
 
-describe('package.json', function () {
-  it('has `bugs`', function () {
-    assert.equal(typeof pkg.bugs, 'string')
+describe('package.json', () => {
+  it('has `bugs`', () => {
+    assert.strictEqual(typeof pkg.bugs, 'string')
   })
 })
 
-describe('package.json and bower.json', function () {
+describe('package.json and bower.json', () => {
   [
     'name',
     'version',
@@ -18,23 +18,23 @@ describe('package.json and bower.json', function () {
     'description',
     'repository',
     'homepage'
-  ].forEach(function (key) {
-    it('have the same ' + key, function () {
-      assert.deepEqual(pkg[key], bower[key])
+  ].forEach(key => {
+    it('have the same ' + key, () => {
+      assert.deepStrictEqual(pkg[key], bower[key])
     })
   })
 
-  it('have the same authors', function () {
-    var pkgAuthors = [pkg.author].concat(pkg.contributors)
-    assert.deepEqual(pkgAuthors, bower.authors)
+  it('have the same authors', () => {
+    const pkgAuthors = [pkg.author].concat(pkg.contributors)
+    assert.deepStrictEqual(pkgAuthors, bower.authors)
   })
 
-  it('are public', function () {
-    assert(!pkg['private'])
-    assert(!bower['private'])
+  it('are public', () => {
+    assert(!pkg.private)
+    assert(!bower.private)
   })
 
-  it('have no dependencies', function () {
+  it('have no dependencies', () => {
     assert(!pkg.dependencies)
     assert(!bower.dependencies)
   })